requirements.txt
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
exceptiongroup==1.3.1
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-telegram-bot==22.5
typing_extensions==4.15.0
Werkzeug==3.1.4
